Moved NativeMachineFrameworkUI into ReBuzzGUI repo (a more suitable place for it, since ReBuzz uses it directly)
Update to project file so that bin and build are not outside the source tree. Also update to submodules

Fixed Pattern Editor selection - as Machine.Name was being used in some places, rather than Machine.Info.Name

Prevent pattern editor machines being shown if they are not in gear.xml.  This is because switching pattern editors uses gear.xml to check the data format supported by the old and new pattern editor.  The editors are unable to exchange data if they are not in gear.xml even if they support the same "data format".   For this reason, any pattern editor not included in gear.xml is NOT shown on the pattern editor combo box.

Renamed ReBuzzPatternXP  to Jeskola Pattern XP, to match entry in gear.xml (See above note)