Added a more effecient mechanism to query what sequences are assocaited with a pattern.  This saves having to loop through ALL sequences and all patterns.
Added pattern play started, play position changed, and play stopped event, needed for NativeMachineFramework to monitor when a pattern is being played.

Removed call of NotifyOfPatternPlaying from PatternXP Tick, into NativeMachineFramework.   The NativeMachineFramework::Tick now works out when to notify of the pattern playing

Replaced Utils::ObjectToInt64() with built-in IntPtr values.  These were not exposed via the interface, so 'getters' have also been added.  This is a more reliable way for the NativeMachineFramework to obtain a unique id for a given pattern, sequence, or machine.